requirements.txt
anthropic>=0.40.0
requests>=2.31.0
PyGithub>=2.1.1
feedparser>=6.0.11
jinja2>=3.1.4
pyyaml>=6.0.1
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
sendgrid>=6.11.0
